plot.dip {diptest}	R Documentation

Plot a dip() Result, i.e., Class "dip" Object

Description

Plot method for "dip" objects, i.e., the result of dip(., full.result=TRUE) or similar.

Note: We may decide to enhance the plot in the future, possibly not entirely back-compatibly.

Usage

## S3 method for class 'dip'
plot(x, do.points = (n < 20),
     colG = "red3", colL = "blue3", colM = "forest green",
     col.points = par("col"), col.hor = col.points,
     doModal = TRUE, doLegend = TRUE, ...)

Arguments

x
an R object of class "dip", i.e., typically the result of dip(., full.result= FF) where FF is TRUE or a string such as "all".

do.points
logical indicating if the ECDF plot should include points; passed to plot.ecdf.

colG, colL, colM
the colors to be used in the graphics for the Greatest convex minorant, the Least concave majorant, and the Modal interval, respectively.

col.points, col.hor
the color of points or horizontal lines, respectively, simply passed to plot.ecdf.

doModal
logical indicating if the modal interval [xL, xU] should be shown.

doLegend
logical indicating if a legend should be shown.

...
further optional arguments, passed to plot.ecdf.

Author(s)

Martin Maechler

See Also

dip, also for examples; plot.ecdf.

[Package diptest version 0.75-7 Index]